+++ App.js
@@ -2,7 +2,7 @@ import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Button, Text, View } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native'; // use command ''
 import { createNativeStackNavigator } from '@react-navigation/native-stack'; // use command 'npm install @react-navigation/native-stack'
-import RestaurantItem from './RestaurantItem';
+import BusinessPage from './BusinessPage';
 import Profile from './Profile'
 import HomeScreen from './HomeScreen';
 
@@ -17,23 +17,23 @@ const Stack = createNativeStackNavigator();
  * @returns Display for mobile device
  */
 export default function App() {
-  return (
-    <View style={styles.container}>
-        <StatusBar style="auto" />
-        <NavigationContainer>
-            <Stack.Navigator initialRouteName="Home">
-                <Stack.Screen name="Home" component={HomeScreen} />
-                <Stack.Screen name="Profile" component={Profile} />
-                <Stack.Screen name="RestaurantItem" component={RestaurantItem} />
-            </Stack.Navigator>
-        </NavigationContainer>
-    </View>
-  );
-} 
+    return (
+        <View style={styles.container}>
+            <StatusBar style="auto" />
+            <NavigationContainer>
+                <Stack.Navigator initialRouteName="Home">
+                    <Stack.Screen name="Home" component={HomeScreen} />
+                    <Stack.Screen name="Profile" component={Profile} />
+                    <Stack.Screen name="BusinessPage" component={BusinessPage} />
+                </Stack.Navigator>
+            </NavigationContainer>
+        </View>
+    );
+}
 
 const styles = StyleSheet.create({
     container: {
-      flex: 1,
-      backgroundColor: '#FFF4BE'
+        flex: 1,
+        backgroundColor: '#FFF4BE'
     }
 })
