+++ babel.config.js
@@ -1,4 +1,3 @@
-
 module.exports = function(api) {
   api.cache(true);
   return {
