+++ package.json
@@ -1,1 +1,1 @@
-{"name":"chowdown","version":"1.0.0","main":"node_modules/expo/AppEntry.js","scripts":{"start":"expo start","android":"expo start --android","ios":"expo start --ios","web":"expo start --web","eject":"expo eject"},"dependencies":{"@react-navigation/native":"^6.0.11","@react-navigation/native-stack":"^6.7.0","expo":"~45.0.0","expo-status-bar":"~1.3.0","react":"17.0.2","react-dom":"17.0.2","react-native":"0.68.2","react-native-stars":"^1.2.2","react-native-web":"0.17.7","react-navigation":"^4.4.4"},"devDependencies":{"@babel/core":"^7.12.9"},"private":true}
+{"name":"chowdown","version":"1.0.0","main":"node_modules/expo/AppEntry.js","scripts":{"start":"expo start","android":"expo start --android","ios":"expo start --ios","web":"expo start --web","eject":"expo eject"},"dependencies":{"@fortawesome/fontawesome-svg-core":"^6.1.2","@fortawesome/free-brands-svg-icons":"^6.1.2","@fortawesome/free-regular-svg-icons":"^6.1.2","@fortawesome/free-solid-svg-icons":"^6.1.2","@fortawesome/react-fontawesome":"^0.2.0","@fortawesome/react-native-fontawesome":"^0.3.0","@primer/octicons-react":"^17.4.0","@react-navigation/native":"^6.0.11","@react-navigation/native-stack":"^6.7.0","babel-plugin-macros":"^3.1.0","expo":"~45.0.0","expo-status-bar":"~1.3.0","react":"17.0.2","react-dom":"17.0.2","react-native":"0.68.2","react-native-stars":"^1.2.2","react-native-svg":"^12.4.3","react-native-vector-icons":"^9.2.0","react-native-web":"0.17.7","react-navigation":"^4.4.4"},"devDependencies":{"@babel/core":"^7.12.9"},"private":true}
